перенос моделей в другой класс
