+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -9,10 +9,13 @@ final class MovieQuizViewController: UIViewController {
     @IBOutlet weak private var noButton: UIButton!
     private var currentQuestionIndex = 0
     private var correctAnswers: Int = 0
+    private let questionsAmount: Int = 10
+    private let questionFactory = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        show(quiz: convert(model: questions[currentQuestionIndex]))
+        show(quiz: convert(model: questionFactory.requestNextQuestion()!))
     }
     
     private func show(quiz step: QuizStepViewModel) {
@@ -34,9 +37,11 @@ final class MovieQuizViewController: UIViewController {
             self.counterLabel.text = "\(self.correctAnswers + 1)/10"
             
             // заново показываем первый вопрос
-            let firstQuestion = questions[self.currentQuestionIndex]
-            let viewModel = self.convert(model: firstQuestion)
-            self.show(quiz: viewModel)
+            if let firstQuestion = self.questionFactory.requestNextQuestion() {
+                self.currentQuestion = firstQuestion
+                let viewModel = self.convert(model: firstQuestion)
+                self.show(quiz: viewModel)
+            }
         }
         alert.addAction(action)
         self.present(alert, animated: true, completion: nil)
@@ -44,13 +49,17 @@ final class MovieQuizViewController: UIViewController {
     
     private func convert(model: QuizQuestion) -> QuizStepViewModel {
         self.textLabel.text = model.text
+        self.currentQuestion = model // тут мы сохроняем текущий вопрос
         return QuizStepViewModel(image: UIImage(named: model.image) ?? UIImage(),
                                  question: model.text,
-                                 questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+                                 questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
     }
     @IBAction func noButton(_ sender: Any) {
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
 
-        if questions[currentQuestionIndex].correctAnswer == false {
+        if currentQuestion.correctAnswer == false {
             showAnswerResult(isCorrect: true)
         } else {
             showAnswerResult(isCorrect: false)
@@ -58,7 +67,12 @@ final class MovieQuizViewController: UIViewController {
     }
     
     @IBAction func yesButton(_ sender: Any) {
-        if questions[currentQuestionIndex].correctAnswer == true {
+        guard let currentQuestion = currentQuestion else {
+            print("error")
+            return
+        }
+        
+        if currentQuestion.correctAnswer == true {
             showAnswerResult(isCorrect: true)
         } else {
             showAnswerResult(isCorrect: false)
@@ -92,83 +106,22 @@ final class MovieQuizViewController: UIViewController {
         yesButton.isEnabled = true
         noButton.isEnabled = true
 
-        if currentQuestionIndex == questions.count - 1 {
+        if currentQuestionIndex == questionsAmount - 1 {
             let text = "Ваш результат: \(correctAnswers) из 10"
             let viewModel = QuizResultsViewModel(title: "Раунд окончен",
                                                  text: text,
                                                  buttonText: "Сыграть еще раз")
             show(quiz: viewModel)
         } else {
-            
             currentQuestionIndex += 1
-            let nextQuestion = questions[self.currentQuestionIndex]
-            let viewModel = convert(model: nextQuestion)
-            self.show(quiz: viewModel)
+            if let nextQuestion = questionFactory.requestNextQuestion() {
+                self.currentQuestion = nextQuestion
+                let viewModel = convert(model: nextQuestion)
+                self.show(quiz: viewModel)
+            }
         }
     }
 }
-
-    // MARK: - Models
-struct QuizQuestion {
-    let image: String
-    let text: String
-    let correctAnswer: Bool
-}
-
-struct QuizStepViewModel {
-    let image: UIImage
-    let question: String
-    let questionNumber: String
-}
-
-struct QuizResultsViewModel {
-    let title: String
-    let text: String
-    let buttonText: String
-}
-    // MARK: - MOK Data
-private let questions: [QuizQuestion] = [
-        QuizQuestion(
-            image: "The Godfather",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Dark Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Kill Bill",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Avengers",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Deadpool",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Green Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Old",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "The Ice Age Adventures of Buck Wild",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Tesla",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Vivarium",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false)
-    ]
 
 /*
  Mock-данные
